requirements.txt
requests==2.32.5
python-dotenv==1.2.1
pandas==2.3.3
scikit-learn==1.7.2